Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.129.0
filelock==3.25.2
google-auth==2.49.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.18
idna==3.11
nodeenv==1.10.0
passlib==1.7.4
pillow==12.1.1
platformdirs==4.9.4
pre_commit==4.5.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-discovery==1.2.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uv==0.11.2
uvicorn==0.40.0
virtualenv==21.2.0
websockets==16.0
